Redesign GlobalNotifier and add fine-grained notifications
Split handling of the admin Realm to a separate AdminRealmManager class, which
no longer uses any of the sync functionality directly and instead just uses the
normal RealmCoordinator API.

Redesign the GlobalNotifier API to pull much more of the functionality required
into it. The binding code no longer needs to subclass GlobalNotifier or
implement helpers to dispatch events to an event loop.

Add basic fine-grained details to the notifications. The notification now
includes the old Realm, new Realm, and which rows of every table changed.

Add some tests for all of this by spawning an in-process sync server.